Create Adjanceny sum array.java
Adjacent Sum Array
Chef had an array 
A
A of length 
N
N of positive integers. He calculated an array 
B
B of 
N
−
1
N−1 integers, where 
B
i
=
A
i
+
A
i
+
1
B 
i
​
 =A 
i
​
 +A 
i+1
​
 .

For example, if 
A
=
[
2
,
5
,
4
]
A=[2,5,4], he got 
B
=
[
2
+
5
,
5
+
4
]
=
[
7
,
9
]
B=[2+5,5+4]=[7,9].

Unfortunately, Chef has forgotten the array 
A
A, and the elements in his sequence 
B
B have been shuffled. Given the shuffled sequence 
B
B, predict the original array 
A
A. If multiple answers exist, all will be accepted. It is guaranteed that the input data is generated in such a way that a valid array 
A
A exists.

Input Format
The first line of input will contain a single integer 
T
T, denoting the number of test cases.
Each test case consists of multiple lines of input.
The first line contains 
N
N - the size of the array 
A
A.
The second line contains 
N
−
1
N−1 integers - the shuffled sequence 
B
B.
Output Format
For each test case, output on a new line 
N
N integers - 
A
1
,
A
2
,
.
.
.
,
A
N
A 
1
​
 ,A 
2
​
 ,...,A 
N
​
  - one of the possible original sequences, 
A
i
>
0
A 
i
​
 >0 for all 
i
i.

If multiple answers are possible, all will be accepted.

Constraints
1
≤
T
≤
1
0
4
1≤T≤10 
4
 
2
≤
N
≤
2
⋅
1
0
5
2≤N≤2⋅10 
5
 
1
≤
B
i
≤
1
0
9
1≤B 
i
​
 ≤10 
9
 
The input is generated such that there exists a valid array 
A
A
The sum of 
N
N over all test cases does not exceed 
2
⋅
1
0
5
2⋅10 
5
 
Sample 1:
Input
Output
2
3
9 7
4
4 4 4
2 5 4
2 2 2 2
Explanation:
Test Case 1 : It is possible that the initial sequence 
A
=
[
2
,
5
,
4
]
A=[2,5,4] because then the constructed initial 
B
B would be 
[
7
,
9
]
[7,9] and then it got shuffled to 
[
9
,
7
]
[9,7]. 
A
=
[
4
,
5
,
2
]
A=[4,5,2] is also a valid initial sequence.

Did you like the problem statement?
80 users found this helpful